requirements.txt
loguru>=0.7.0
numpy>=1.26.0
pandas>=2.0.0
pypdf>=3.17.0
python-dotenv==1.0.0
pyyaml>=6.0.1
requests==2.26.0
